feat(puck): rename FilePickerField to ImagePickerField
- Rename FilePickerField → ImagePickerField, SelectedFile → SelectedImage
- Better reflects image-centric implementation (uses <img>, has alt field)
- Remove aspectRatio from demo Image component (simplify)
- Add optional filename segment to file URLs for SEO
  - CMS: /admin/files/{table}/{column}/{id}[/{filename}]
  - Demo: /files/media/{id}[/{filename}]
- Update README documentation
