+++ apps/demo/components.tsx
@@ -10,251 +10,11 @@
  */
 
 import { DropZone, React } from '@hotsauce/plugins/puck/client/globals';
+import {
+  ImagePickerField,
+  type SelectedImage,
+} from '@hotsauce/plugins/puck/fields';
 import type { ComponentConfig, PuckProps } from '@hotsauce/plugins/puck/types';
-
-// ============================================================================
-// Media Reference type (matches CMS media table structure)
-// ============================================================================
-
-/**
- * Reference to a media record stored in the CMS.
- * Stores only identifiers — URLs are constructed at render time.
- */
-type MediaReference = {
-  /** Primary key from the media table */
-  id: number;
-  /** Table name (e.g. 'media', 'assets') */
-  table: string;
-  /** File column name (e.g. 'file', 'image') */
-  column: string;
-  /** Alt text seeded from the media record (can be overridden per usage) */
-  alt?: string;
-  /** Original filename (display only) */
-  filename?: string;
-};
-
-// ============================================================================
-// ImagePicker - Custom field component for selecting images from CMS media
-// ============================================================================
-
-/**
- * Custom Puck field component that opens a media picker dialog.
- *
- * Opens the CMS grid view in an iframe with `?picker=true` mode.
- * When the user clicks an image, the picker posts a `cms:media-selected`
- * message containing the record data and resolved URL.
- *
- * @example
- * ```tsx
- * // In a Puck component config:
- * fields: {
- *   image: {
- *     type: 'custom',
- *     label: 'Image',
- *     render: ({ value, onChange }) => (
- *       <ImagePickerField value={value} onChange={onChange} />
- *     ),
- *   },
- * }
- * ```
- */
-function ImagePickerField({
-  value,
-  onChange,
-  basePath = '/admin',
-  table = 'media',
-  column = 'file',
-  altField = 'alt',
-}: {
-  value: MediaReference | null;
-  onChange: (value: MediaReference | null) => void;
-  /** Base path where the CMS is mounted (default: '/admin') */
-  basePath?: string;
-  /** Table name to pick from (default: 'media') */
-  table?: string;
-  /** File column name on the table (default: 'file') */
-  column?: string;
-  /** Column name for alt text on the record (default: 'alt') */
-  altField?: string;
-}) {
-  const [isOpen, setIsOpen] = React.useState(false);
-  const dialogRef = React.useRef<HTMLDialogElement>(null);
-
-  // Handle postMessage from picker iframe
-  React.useEffect(() => {
-    function handleMessage(event: MessageEvent) {
-      if (event.data?.type === 'cms:media-selected') {
-        const record = event.data.record;
-        const file = record?.[column];
-
-        if (record?.id != null) {
-          onChange({
-            id: record.id,
-            table: event.data.table || table,
-            column,
-            alt: (altField && record[altField]) || '',
-            filename: file?.filename || '',
-          });
-        }
-        setIsOpen(false);
-        dialogRef.current?.close();
-      }
-    }
-
-    if (isOpen) {
-      // deno-lint-ignore no-window no-window-prefix
-      window.addEventListener('message', handleMessage);
-      // deno-lint-ignore no-window no-window-prefix
-      return () => window.removeEventListener('message', handleMessage);
-    }
-  }, [isOpen, onChange]);
-
-  const openPicker = () => {
-    setIsOpen(true);
-    dialogRef.current?.showModal();
-  };
-
-  const clearSelection = () => {
-    onChange(null);
-  };
-
-  return (
-    <div style={{ display: 'flex', flexDirection: 'column', gap: '8px' }}>
-      {value?.id
-        ? (
-          <div style={{ position: 'relative' }}>
-            <img
-              src={`${basePath}/files/${value.table}/${value.column}/${value.id}`}
-              alt={value.alt || ''}
-              style={{
-                width: '100%',
-                maxHeight: '200px',
-                objectFit: 'contain',
-                borderRadius: '4px',
-                backgroundColor: '#f0f0f0',
-              }}
-            />
-            <div style={{ marginTop: '4px', fontSize: '12px', color: '#666' }}>
-              {value.filename || `ID: ${value.id}`}
-            </div>
-          </div>
-        )
-        : (
-          <div
-            style={{
-              width: '100%',
-              height: '100px',
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              backgroundColor: '#f5f5f5',
-              borderRadius: '4px',
-              color: '#999',
-            }}
-          >
-            No image selected
-          </div>
-        )}
-
-      <div style={{ display: 'flex', gap: '8px' }}>
-        <button
-          type='button'
-          onClick={openPicker}
-          style={{
-            flex: 1,
-            padding: '8px 12px',
-            backgroundColor: '#0066cc',
-            color: 'white',
-            border: 'none',
-            borderRadius: '4px',
-            cursor: 'pointer',
-            fontSize: '14px',
-          }}
-        >
-          {value ? 'Change Image' : 'Pick Image'}
-        </button>
-        {value && (
-          <button
-            type='button'
-            onClick={clearSelection}
-            style={{
-              padding: '8px 12px',
-              backgroundColor: '#dc3545',
-              color: 'white',
-              border: 'none',
-              borderRadius: '4px',
-              cursor: 'pointer',
-              fontSize: '14px',
-            }}
-          >
-            Remove
-          </button>
-        )}
-      </div>
-
-      <dialog
-        ref={dialogRef}
-        style={{
-          width: '90vw',
-          maxWidth: '900px',
-          height: '80vh',
-          padding: 0,
-          border: 'none',
-          borderRadius: '8px',
-          boxShadow: '0 4px 20px rgba(0,0,0,0.3)',
-        }}
-        onClose={() => setIsOpen(false)}
-      >
-        <div
-          style={{
-            display: 'flex',
-            flexDirection: 'column',
-            height: '100%',
-          }}
-        >
-          <div
-            style={{
-              display: 'flex',
-              justifyContent: 'space-between',
-              alignItems: 'center',
-              padding: '12px 16px',
-              borderBottom: '1px solid #eee',
-              backgroundColor: '#f9f9f9',
-            }}
-          >
-            <strong>Select Image</strong>
-            <button
-              type='button'
-              onClick={() => {
-                setIsOpen(false);
-                dialogRef.current?.close();
-              }}
-              style={{
-                padding: '4px 12px',
-                border: '1px solid #ccc',
-                borderRadius: '4px',
-                backgroundColor: 'white',
-                cursor: 'pointer',
-              }}
-            >
-              Cancel
-            </button>
-          </div>
-          {isOpen && (
-            <iframe
-              src={`${basePath}/${table}?picker=true`}
-              style={{
-                flex: 1,
-                width: '100%',
-                border: 'none',
-              }}
-            />
-          )}
-        </div>
-      </dialog>
-    </div>
-  );
-}
 
 // ============================================================================
 // Heading
@@ -391,7 +151,7 @@ const Image: ComponentConfig = {
       label: 'Image',
       render: ({ value, onChange }) => (
         <ImagePickerField
-          value={value as MediaReference | null}
+          value={value as SelectedImage | null}
           onChange={onChange}
         />
       ),
@@ -400,50 +160,31 @@ const Image: ComponentConfig = {
       type: 'text',
       label: 'Alt Text',
     },
-    aspectRatio: {
-      type: 'select',
-      label: 'Aspect Ratio',
-      options: [
-        { label: 'Auto', value: 'auto' },
-        { label: '16:9', value: '16:9' },
-        { label: '4:3', value: '4:3' },
-        { label: '1:1', value: '1:1' },
-      ],
-    },
   },
   defaultProps: {
     media: null,
     alt: '',
-    aspectRatio: 'auto',
   },
-  resolveFields: (data) => {
-    const m = data.props.media as MediaReference | null;
+  resolveFields: (data, { fields }) => {
+    const m = data.props.media as SelectedImage | null;
     return {
-      media: Image.fields!.media,
+      ...fields,
       alt: {
         type: 'text' as const,
         label: 'Alt Text',
         placeholder: m?.alt || 'Describe this image…',
       },
-      aspectRatio: Image.fields!.aspectRatio,
     };
   },
-  render: ({ media, alt, aspectRatio, puck }) => {
-    const m = media as MediaReference | null;
+  render: ({ media, alt, puck }) => {
+    const m = media as SelectedImage | null;
     // Puck alt field takes precedence; fall back to alt seeded from media record
     const altText = (alt as string) || m?.alt || '';
-    const ratioMod = aspectRatio === '16:9'
-      ? 'image--ratio-16-9'
-      : aspectRatio === '4:3'
-      ? 'image--ratio-4-3'
-      : aspectRatio === '1:1'
-      ? 'image--ratio-1-1'
-      : 'image--ratio-auto';
 
     if (!m?.id) {
       return (
         <div
-          className={`image image--placeholder ${ratioMod}`}
+          className='image image--placeholder'
           style={{
             display: 'flex',
             alignItems: 'center',
@@ -458,19 +199,15 @@ const Image: ComponentConfig = {
       );
     }
 
-    // In the editor, use the private CMS URL; on the public site, use /files/:table/:id
+    // In the editor, use the private CMS URL; on the public site, use /files/:table/:id/:filename
     const isEditing = (puck as { isEditing?: boolean })?.isEditing;
     const src = isEditing
       ? `/admin/files/${m.table}/${m.column}/${m.id}`
-      : `/files/${m.table}/${m.id}`;
+      : `/files/${m.table}/${m.id}${
+        m.filename ? `/${encodeURIComponent(m.filename)}` : ''
+      }`;
 
-    return (
-      <img
-        src={src}
-        alt={altText}
-        className={`image ${ratioMod}`}
-      />
-    );
+    return <img src={src} alt={altText} className='image' />;
   },
 };
