+++ apps/demo/site/static/components.css
@@ -99,23 +99,6 @@
   display: block;
 }
 
-/* Aspect Ratio Modifiers */
-.image--ratio-auto {
-  aspect-ratio: auto;
-}
-
-.image--ratio-16-9 {
-  aspect-ratio: 16 / 9;
-}
-
-.image--ratio-4-3 {
-  aspect-ratio: 4 / 3;
-}
-
-.image--ratio-1-1 {
-  aspect-ratio: 1 / 1;
-}
-
 /* ==========================================================================
    Section Block
    ========================================================================== */
