+++ apps/demo/site/static/styles.css
@@ -7,7 +7,11 @@
   --color-border: #e5e5e5;
   --color-card-bg: #fff;
   --font-sans:
-    system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
+    system-ui,
+    -apple-system,
+    BlinkMacSystemFont,
+    'Segoe UI',
+    Roboto,
     sans-serif;
   --font-serif: Georgia, 'Times New Roman', serif;
   --max-width: 1100px;
